GH-1223 polishing
