+++ buildSrc/src/main/kotlin/Dependencies.kt
@@ -7,6 +7,5 @@ object Dependencies {
     const val jspecifyVersion = "1.0.0"
     const val jsonUnitAssertJVersion = "2.38.0"
     const val jsonwebtokenVersion = "0.12.6"
-    const val instancioVersion = "5.5.1"
     const val junitPlatformLauncherVersion = "1.12.2"
 }
