+++ master/build.gradle.kts
@@ -17,6 +17,7 @@ val springDocSwaggerVersion = "3.0.3"
 val sqliteVersion = "3.53.1.0"
 val nimbusJoseJwt ="10.9.1"
 val jmesPathVersion = "0.6.0"
+val instancioVersion = "5.5.1"
 
 // Explicitly specified versions for security reasons (i.e. using some specific patch versions)
 val postgresqlVersion = "42.7.11"
@@ -92,6 +93,7 @@ dependencies {
     testImplementation("com.squareup.okhttp3:mockwebserver")
     testImplementation("com.squareup.okhttp3:okhttp")
     testImplementation("digital.pragmatech.testing:spring-test-profiler:0.1.1")
+    testImplementation("org.instancio:instancio-core:${instancioVersion}")
 
     // annotation processor
     annotationProcessor("org.springframework.boot:spring-boot-configuration-processor:$springBootVersion")
