+++ sbs/axelix-spring-boot-3-starter/src/test/java/com/axelixlabs/axelix/sbs/spring/core/transactions/TransactionMonitoringEndpointTest.java
@@ -33,23 +33,23 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 
-import ch.qos.logback.classic.LoggerContext;
 import com.jayway.jsonpath.JsonPath;
 import io.micrometer.core.instrument.Counter;
 import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.Timer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.slf4j.LoggerFactory;
 
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Lazy;
+import org.springframework.context.event.EventListener;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -642,18 +642,9 @@ class TransactionMonitoringEndpointTest {
             return new DefaultAxelixMetricsPublisher(meterRegistry);
         }
 
-        @Bean
-        public InMemoryPaginationAppender inMemoryPaginationAppender() {
-            LoggerContext context = (LoggerContext) LoggerFactory.getILoggerFactory();
-
-            InMemoryPaginationAppender appender = new InMemoryPaginationAppender();
-            appender.setContext(context);
-            appender.start();
-
-            // Hibernate 6+ (Spring Boot 3)
-            context.getLogger("org.hibernate.orm.query").addAppender(appender);
-
-            return appender;
+        @EventListener(ApplicationReadyEvent.class)
+        public void registerAppender() {
+            new LogbackInMemoryPaginationAppenderRegistrar().register();
         }
     }
